Basic joystick control
This is used to control the motor with a joystick and verify that everything is working correctly (like the potentiometer and current values). Be careful in this code the movement isn't blocked by the potentiometer position so it is possible to go out of bounds and possibly break the motor or some parts of the gripper. Proceed with care.